+++ input_harmonization.cpp
@@ -97,7 +97,9 @@ void Input::DoAdaptExternalities()
 
     GleanedLedgerType_ =
         static_cast<mcenum_ledger_type>
-            (database_->Query(DB_LedgerType)
+            (static_cast<int>
+                (database_->Query(DB_LedgerType)
+                )
             );
 }
 
@@ -276,7 +278,9 @@ void Input::DoHarmonize()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (database_->Query(DB_AgeLastOrNearest)
+            (static_cast<int>
+                (database_->Query(DB_AgeLastOrNearest)
+                )
             );
     DateOfBirth.minimum_and_maximum
         (minimum_birthdate(IssueAge.maximum(), EffectiveDate.value(), alb_anb)
@@ -916,7 +920,9 @@ void Input::DoTransmogrify()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (database_->Query(DB_AgeLastOrNearest)
+            (static_cast<int>
+                (database_->Query(DB_AgeLastOrNearest)
+                )
             );
 
     int apparent_age = attained_age
